… cmd/oteltestbedcol/go.sum
… connector/countconnector/go.sum
… processor/attributesprocessor/go.sum
… processor/spanprocessor/go.sum