**Description:**
This PR adds a bridge between `filterspan.NewSkipExpr` and `filterottl.NewBoolExprForSpan`.  With this change, any component using `filterspan.NewSkipExpr` will start using OTTL behind the scenes.  

**Link to tracking Issue:**
Related to https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/18643 
Related to https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/18642

**Testing:**
All unit tests from any component using filterspan are now testing this bridge.