@TylerHelmuth 2023-04-26
@dmitryax here is a quick benchmark comparison between `filterspan` using `filtermatcher`/`filterset` and `filterspan` using `filterottl`.  I only wrote up 1 static and 1 regex case so far as I think it shows the general change in performance, but I can add more tests for more scenarios as well.

The tests against `filtermatcher`/`filterset`, were run from [this branch](https://github.com/TylerHelmuth/opentelemetry-collector-contrib/blob/1772ead3da5ca093aadbfbe1ab82056bff72eb9d/internal/filter/filterspan/filterspan_test.go#L318).

**`filtermatcher`/`filterset` results:**
```bash
❯ go test -bench=. -run=notests --tags="" -benchmem github.com/open-telemetry/opentelemetry-collector-contrib/internal/filter/filterspan
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/filter/filterspan
BenchmarkFilterspan_NewSkipExpr/service_name_match_regexp-10             4606274               247.7 ns/op             0 B/op          0 allocs/op
BenchmarkFilterspan_NewSkipExpr/service_name_match_static-10             6298548               189.9 ns/op             0 B/op          0 allocs/op
```

**`filterottl` results**
```bash
❯ go test -bench=. -run=notests --tags="" -benchmem github.com/open-telemetry/opentelemetry-collector-contrib/internal/filter/filterspan
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/filter/filterspan
BenchmarkFilterspan_NewSkipExpr/service_name_match_regexp-10             3404667               336.1 ns/op            32 B/op          2 allocs/op
BenchmarkFilterspan_NewSkipExpr/service_name_match_static-10             4759767               252.1 ns/op            16 B/op          1 allocs/op
```

The results show we do take a performance hit using OTTL.  Both cases gain an allocation when grabbing a value from the resource attributes and the regex case gets an extra allocation since it uses [StringLikeGetter](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/167c2f0432a59156bef7fafd564e506802a73184/pkg/ottl/expression.go#L175) to try and convert the value to a string.

Personally I think the hit to performance is worth the flexibility and maintainability we gain by centralizing filtering around OTTL.  @open-telemetry/collector-contrib-approvers what do you think?