@TylerHelmuth 2023-05-01
With the new feature allowing converters to be used as constant booleans (https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/20911), we have narrowed the performance gap slightly.

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/filter/filterspan
                                                    │   old.txt   │               new.txt               │
                                                    │   sec/op    │   sec/op     vs base                │
Filterspan_NewSkipExpr/service_name_match_regexp-10   220.3n ± 1%   292.4n ± 1%  +32.75% (p=0.000 n=10)
Filterspan_NewSkipExpr/service_name_match_static-10   175.5n ± 0%   223.4n ± 0%  +27.29% (p=0.000 n=10)
geomean                                               196.6n        255.6n       +29.99%

                                                    │   old.txt   │           new.txt            │
                                                    │    B/op     │    B/op     vs base          │
Filterspan_NewSkipExpr/service_name_match_regexp-10   0.00 ± 0%     32.00 ± 0%  ? (p=0.000 n=10)
Filterspan_NewSkipExpr/service_name_match_static-10   0.00 ± 0%     16.00 ± 0%  ? (p=0.000 n=10)
geomean                                                         ¹   22.63       ?
¹ summaries must be >0 to compute geomean

                                                    │   old.txt    │           new.txt            │
                                                    │  allocs/op   │ allocs/op   vs base          │
Filterspan_NewSkipExpr/service_name_match_regexp-10   0.000 ± 0%     2.000 ± 0%  ? (p=0.000 n=10)
Filterspan_NewSkipExpr/service_name_match_static-10   0.000 ± 0%     1.000 ± 0%  ? (p=0.000 n=10)
geomean                                                          ¹   1.414       ?
¹ summaries must be >0 to compute geomean
```

The current implementation doesn't check if it needs to use `()` for groupings, it uses them whenever they could possibly be needed. I will experiment and see if writing the most efficient statement (least amount of characters), results in the most performant statement due to the decrease in the boolean expression chain.