@TylerHelmuth 2023-05-03
Continued increase from removing unneeded parentheses:

```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/internal/filter/filterspan
                                                    │   old.txt   │               new.txt               │
                                                    │   sec/op    │   sec/op     vs base                │
Filterspan_NewSkipExpr/service_name_match_regexp-10   223.1n ± 1%   290.9n ± 1%  +30.37% (p=0.000 n=10)
Filterspan_NewSkipExpr/service_name_match_static-10   176.2n ± 1%   222.1n ± 0%  +26.05% (p=0.000 n=10)
geomean                                               198.3n        254.2n       +28.19%

                                                    │   old.txt   │           new.txt            │
                                                    │    B/op     │    B/op     vs base          │
Filterspan_NewSkipExpr/service_name_match_regexp-10   0.00 ± 0%     32.00 ± 0%  ? (p=0.000 n=10)
Filterspan_NewSkipExpr/service_name_match_static-10   0.00 ± 0%     16.00 ± 0%  ? (p=0.000 n=10)
geomean                                                         ¹   22.63       ?
¹ summaries must be >0 to compute geomean

                                                    │   old.txt    │           new.txt            │
                                                    │  allocs/op   │ allocs/op   vs base          │
Filterspan_NewSkipExpr/service_name_match_regexp-10   0.000 ± 0%     2.000 ± 0%  ? (p=0.000 n=10)
Filterspan_NewSkipExpr/service_name_match_static-10   0.000 ± 0%     1.000 ± 0%  ? (p=0.000 n=10)
geomean                                                          ¹   1.414       ?
¹ summaries must be >0 to compute geomean
```